pyproject.toml
[tool.poetry]
name = "tfd-sdk"
version = "0.1.0"
description = "Collection of tools and sdks to work with Nexon's TFD API"
authors = ["Erik Karlsson <erikdkarlsson8@gmail.com>"]
readme = "README.md"
keywords = ["OpenAPI", "OpenAPI-Generator", "TFD Metadata API", "TFD Recommendation API", "TFD User API"]
packages = [
    #    {include = "tfd_sdk/clients/meta"},
    #    {include = "tfd_sdk/clients/user"},
    #    {include = "tfd_sdk/clients/module_recommendation"},
]

include = ["meta_client/py.typed"]

[tool.poetry.dependencies]
pydantic = "^2.8.2"
python = ">=3.11"
python-dateutil = ">=2.8.2"
pyyaml = "^6.0.2"
typing-extensions = ">=4.7.1"
urllib3 = ">= 1.25.3"

[tool.poetry.group.dev.dependencies]
black = "^24.8.0"
flake8 = ">=4.0.0"
ipython = { version = "^8.26.0", optional = true }
mypy = "^1.11.1"
pytest = ">=7.2.1"
python-dotenv = "^1.0.1"
setuptools = "^72.1.0"
tox = ">=3.9.0"
types-python-dateutil = ">=2.8.19.14"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
format-api-spec = "tfd_sdk.util.format_api_spec:main"
[tool.mypy]
files = [
    "meta_client",
    "user_client",
    "recommendation_client",
    #"test",  # auto-generated tests
    "tests", # handwritten tests
]

# List from: https://mypy.readthedocs.io/en/stable/existing_code.html#introduce-stricter-options
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true

## Getting these passing should be easy
strict_equality = true
strict_concatenate = true

## Strongly recommend enabling this one as soon as you can
check_untyped_defs = true

## These shouldn't be too much additional work, but may be tricky to
## get passing if you use a lot of untyped libraries
disallow_subclassing_any = true
disallow_untyped_decorators = true
disallow_any_generics = true

### These next few are various gradations of forcing use of type annotations
#disallow_untyped_calls = true
#disallow_incomplete_defs = true
#disallow_untyped_defs = true
#
### This one isn't too hard to get passing, but return on investment is lower
#no_implicit_reexport = true
#
### This one can be tricky to get passing if you use a lot of untyped libraries
#warn_return_any = true
